refactor IndividualFlags component
